Extract card and popup to Recipe.js.
